requirements.txt
discord==1.0.1
discord.py==2.3.0
python-dateutil==2.8.2
python-dotenv==1.0.0
